requirements/requirements.txt
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
tensorflow-macos==2.15.0
tensorflow-metal==0.8.0
opencv-python
numpy
pandas
matplotlib
pillow
tqdm
fastapi==0.121.0
uvicorn==0.38.0
editdistance
jiwer
nltk
pydantic==2.12.3
rich
tensorboard

